# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSB HRD API",
        "description": "API documentation for HRD modules in SSB Project",
        "contact": {
            "email": "admin@ssb.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://hrd.ptssb.my.id/api",
            "description": "SSB API Server"
        }
    ],
    "paths": {
...
        "/media/photo/{filename}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Get foto karyawan",
                "description": "Mengambil file foto karyawan dari folder storage/app/public/hrd/photo. Memerlukan service token.",
                "operationId": "ed844e54cb4aa11f7ea73fabeba94adf",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Nama file foto (contoh: foto_001.jpg)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "foto_001.jpg"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File binary (image)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "File tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/media/memo-internal/{filename}": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Get file memo internal",
                "description": "Mengambil file memo internal dari folder storage/app/public/memo_internal. Memerlukan service token.",
                "operationId": "c94bc76a66374eea5de7d8d65e55940f",
                "parameters": [
                    {
                        "name": "filename",
                        "in": "path",
                        "description": "Nama file memo (contoh: memo_001.pdf)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "memo_001.pdf"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File binary (pdf/image)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "File tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/service-token": {
            "get": {
                "tags": [
                    "Service Token"
                ],
                "summary": "List semua service token aktif",
                "description": "Menampilkan daftar semua permanent service token yang telah digenerate untuk aplikasi eksternal.",
                "operationId": "9ef82984aa9bae478b35cc50a73faa21",
                "responses": {
                    "200": {
                        "description": "Daftar service token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "app_name": {
                                                        "type": "string",
                                                        "example": "Aplikasi HRD External"
                                                    },
                                                    "nik": {
                                                        "type": "string",
                                                        "example": "001"
                                                    },
                                                    "abilities": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string"
                                                        },
                                                        "example": [
                                                            "*"
                                                        ]
                                                    },
                                                    "last_used_at": {
                                                        "type": "string",
                                                        "example": "2026-05-15 08:00:00",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2026-05-15 07:00:00"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Service Token"
                ],
                "summary": "Generate permanent service token",
                "description": "Membuat token permanen untuk aplikasi eksternal. Token hanya ditampilkan sekali - simpan segera.",
                "operationId": "bef9668439bb0f17a19a40e8d7d124ff",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "app_name"
                                ],
                                "properties": {
                                    "nik": {
                                        "description": "NIK user / service account",
                                        "type": "string",
                                        "example": "001"
                                    },
                                    "app_name": {
                                        "description": "Nama aplikasi yang memakai token",
                                        "type": "string",
                                        "example": "Aplikasi HRD External"
                                    },
                                    "abilities": {
                                        "description": "Hak akses token. Default [*] berarti semua akses.",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "*"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Token berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service token berhasil dibuat. Simpan token ini - tidak akan ditampilkan lagi."
                                        },
                                        "token_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "app_name": {
                                            "type": "string",
                                            "example": "Aplikasi HRD External"
                                        },
                                        "nik": {
                                            "type": "string",
                                            "example": "001"
                                        },
                                        "abilities": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "*"
                                            ]
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "5|AbCdEfGhIj..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/service-token/{id}": {
            "delete": {
                "tags": [
                    "Service Token"
                ],
                "summary": "Cabut (revoke) service token",
                "description": "Menghapus permanent service token berdasarkan ID. Aplikasi yang memakai token ini tidak akan bisa akses API lagi.",
                "operationId": "a65affca027a63bff14a67324185899e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID token (dari kolom id tabel personal_access_tokens)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token berhasil dicabut",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Token ID 5 berhasil dicabut."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Token tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}